precision-oncology-ml-sandbox/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "precision-oncology-ml-sandbox"
version = "0.1.0"
description = "Deterministic, end-to-end ML sandbox for precision oncology across genomics, imaging, NLP, and RL."
readme = "README.md"
requires-python = ">=3.11,<3.12"
license = {text = "MIT"}
authors = [{ name = "Your Name" }]
dependencies = [
  "numpy>=1.26",
  "pandas>=2.2",
  "scikit-learn>=1.5",
  "hydra-core>=1.3",
  "mlflow>=2.16",
  "Pillow>=10.3",
  "matplotlib>=3.8",
  "typer>=0.12",
  "rich>=13.8",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "pytest-cov>=5.0",
  "ruff>=0.6",
  "black>=24.8",
  "mypy>=1.11",
  "pre-commit>=3.8",
]
torch = [
  "torch>=2.3",
]
sentencetransformers = [
  "sentence-transformers>=3.0",
]

[project.urls]
homepage = "https://github.com/your-org/precision-oncology-ml-sandbox"

[project.scripts]
pon-generate-genomics = "pon_ml.data.gen_genomics:main"
pon-generate-imaging = "pon_ml.data.gen_imaging:main"
pon-generate-nlp = "pon_ml.data.gen_nlp:main"
pon-generate-ehr = "pon_ml.data.gen_ehr:main"
pon-train-genomics = "pon_ml.genomics.train:main"
pon-train-imaging = "pon_ml.imaging.train:main"
pon-nlp-qa = "pon_ml.nlp.rag:cli"
pon-rl-train = "pon_ml.rl.train:main"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
lint.select = ["E","F","I","B","C90","UP","S","N","A","C4","RUF"]
lint.ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q --cov=src/pon_ml --cov-report=term-missing"

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
